Add support for curl config file for each entry

+++ url-checker.sh
@@ -12,9 +12,10 @@ then
       then
         # The format of the file is: 404 URL, so use cut to split on space
         # the first field will be the error code and the second will be the 
-        # URL pattern.
+        # URL pattern. The optional third field will be the optional curl config file.
         HTTP_CODE=`echo $p | cut -f 1 -d " "`
         URL=`echo $p | cut -f 2 -d " "`
+        CONFIG_FILE=`echo $p | cut -f 3 -d " "`
 
         # Have to use the dread eval() here since brace expansion occurs before 
         # parameter expansion. Be careful that the URL list in the file 
@@ -23,7 +24,12 @@ then
         for OUTPUT_URL in $(eval echo "$URL");
         do 
           # Generate the curl command to output the HTTP response code
-          CURL=`curl -sL -w "%{http_code}" $OUTPUT_URL -o /dev/null`
+          if [ -f $CONFIG_FILE ] 
+          then
+            CURL=`curl -sL -K $CONFIG_FILE -w "%{http_code}" $OUTPUT_URL -o /dev/null`
+          else
+            CURL=`curl -sL -w "%{http_code}" $OUTPUT_URL -o /dev/null`
+          fi
 
           if [ $CURL != $HTTP_CODE ]
           then
